rename point
