+++ src/EventTypeDefs.ts
@@ -1,5 +1,5 @@
 import type { Control } from './controls/Control';
-import type { Point } from './point.class';
+import type { Point } from './Point';
 import type { FabricObject } from './shapes/Object/FabricObject';
 import type { FabricObject as StaticFabricObject } from './shapes/Object/Object';
 import type { FabricObjectSVGExportMixin } from './shapes/Object/FabricObjectSVGExportMixin';
