+++ src/brushes/base_brush.class.ts
@@ -1,5 +1,5 @@
 import { Color } from '../color/Color';
-import type { Point } from '../point.class';
+import type { Point } from '../Point';
 import { TEvent } from '../EventTypeDefs';
 import type { Shadow } from '../shadow.class';
 import type { Canvas } from '../canvas/canvas_events';
